+++ web/src/js/__tests__/ducks/tutils.ts
@@ -79,6 +79,7 @@ export const testState: RootState = {
     options: defaultOptions,
     flows: {
         selected: [tflow1.id],
+        isMultipleFlowsSelection: false,
         byId: {
             [tflow0.id]: tflow0,
             [tflow1.id]: tflow1,