+++ web/src/js/components/Header/FlowMenu.tsx
@@ -11,6 +11,8 @@ import {
     replay as replayFlow,
     resume as resumeFlow,
     revert as revertFlow,
+    removeMultiple as removeMultipleFlows,
+    select as selectFlow,
 } from "../../ducks/flows";
 import Dropdown, { MenuItem } from "../common/Dropdown";
 import { copy } from "../../flow/export";
@@ -24,6 +26,8 @@ export default function FlowMenu(): JSX.Element {
         (state) => state.flows.byId[state.flows.selected[0]],
     );
 
+    const selectedFlows = useAppSelector((state) => state.flows.selected);
+
     if (!flow) return <div />;
     return (
         <div className="flow-menu">
@@ -34,7 +38,9 @@ export default function FlowMenu(): JSX.Element {
                             title="[r]eplay flow"
                             icon="fa-repeat text-primary"
                             onClick={() => dispatch(replayFlow(flow))}
-                            disabled={!canReplay(flow)}
+                            disabled={
+                                !canReplay(flow) || selectedFlows.length > 1
+                            }
                         >
                             Replay
                         </Button>
@@ -42,11 +48,16 @@ export default function FlowMenu(): JSX.Element {
                             title="[D]uplicate flow"
                             icon="fa-copy text-info"
                             onClick={() => dispatch(duplicateFlow(flow))}
+                            disabled={selectedFlows.length > 1}
                         >
                             Duplicate
                         </Button>
                         <Button
-                            disabled={!flow || !flow.modified}
+                            disabled={
+                                !flow ||
+                                !flow.modified ||
+                                selectedFlows.length > 1
+                            }
                             title="revert changes to flow [V]"
                             icon="fa-history text-warning"
                             onClick={() => dispatch(revertFlow(flow))}
@@ -56,10 +67,20 @@ export default function FlowMenu(): JSX.Element {
                         <Button
                             title="[d]elete flow"
                             icon="fa-trash text-danger"
-                            onClick={() => dispatch(removeFlow(flow))}
+                            onClick={() => {
+                                if (selectedFlows.length > 1) {
+                                    dispatch(
+                                        removeMultipleFlows(selectedFlows),
+                                    );
+                                    dispatch(selectFlow(undefined)); // clear flow selection
+                                } else {
+                                    dispatch(removeFlow(flow));
+                                }
+                            }}
                         >
                             Delete
                         </Button>
+
                         <MarkButton flow={flow} />
                     </div>
                     <div className="menu-legend">Flow Modification</div>
@@ -78,7 +99,11 @@ export default function FlowMenu(): JSX.Element {
                 <div className="menu-group">
                     <div className="menu-content">
                         <Button
-                            disabled={!flow || !flow.intercepted}
+                            disabled={
+                                !flow ||
+                                !flow.intercepted ||
+                                selectedFlows.length > 1
+                            }
                             title="[a]ccept intercepted flow"
                             icon="fa-play text-success"
                             onClick={() => dispatch(resumeFlow(flow))}
@@ -86,7 +111,11 @@ export default function FlowMenu(): JSX.Element {
                             Resume
                         </Button>
                         <Button
-                            disabled={!flow || !flow.intercepted}
+                            disabled={
+                                !flow ||
+                                !flow.intercepted ||
+                                selectedFlows.length > 1
+                            }
                             title="kill intercepted flow [x]"
                             icon="fa-times text-danger"
                             onClick={() => dispatch(killFlow(flow))}
@@ -108,6 +137,9 @@ const openInNewTab = (url) => {
 };
 
 function DownloadButton({ flow }: { flow: Flow }) {
+    const selectedFlows = useAppSelector((state) => state.flows.selected);
+    if (selectedFlows.length > 1) return null;
+
     if (flow.type !== "http")
         return (
             <Button icon="fa-download" onClick={() => 0} disabled>
@@ -178,6 +210,7 @@ function DownloadButton({ flow }: { flow: Flow }) {
 }
 
 function ExportButton({ flow }: { flow: Flow }) {
+    const selectedFlows = useAppSelector((state) => state.flows.selected);
     return (
         <Dropdown
             className=""
@@ -186,7 +219,7 @@ function ExportButton({ flow }: { flow: Flow }) {
                     title="Export flow."
                     icon="fa-clone"
                     onClick={() => 1}
-                    disabled={flow.type !== "http"}
+                    disabled={flow.type !== "http" || selectedFlows.length > 1}
                 >
                     Export▾
                 </Button>
@@ -222,6 +255,7 @@ const markers = {
 
 function MarkButton({ flow }: { flow: Flow }) {
     const dispatch = useAppDispatch();
+    const selectedFlows = useAppSelector((state) => state.flows.selected);
     return (
         <Dropdown
             className=""
@@ -230,6 +264,7 @@ function MarkButton({ flow }: { flow: Flow }) {
                     title="mark flow"
                     icon="fa-paint-brush text-success"
                     onClick={() => 1}
+                    disabled={selectedFlows.length > 1}
                 >
                     Mark▾
                 </Button>