+++ CHANGELOG.md
@@ -24,6 +24,8 @@
   ([#7286](https://github.com/mitmproxy/mitmproxy/pull/7286), @lukant)
 - Fix a crash when handling corrupted compressed body in savehar addon and its tests.
   ([#7320](https://github.com/mitmproxy/mitmproxy/pull/7320), @8192bytes)
+- Add multiple flows selection to mitmweb.
+  ([#7319](https://github.com/mitmproxy/mitmproxy/pull/7319), @lups2000)
 
 ## 02 October 2024: mitmproxy 11.0.0